+++ src/messagingsession/DefaultMessagingSession.ts
@@ -108,12 +108,32 @@ export default class DefaultMessagingSession implements MessagingSession {
 
   private async startConnecting(reconnecting: boolean): Promise<void> {
     let endpointUrl = this.configuration.endpointUrl;
+
+    // Moving this reconnect logic can potentially result into an infinite reconnect loop on errors.
+    // Check https://github.com/aws/amazon-chime-sdk-js/issues/2372 for details.
+    if (!reconnecting) {
+      this.reconnectController.reset();
+    }
+    if (this.reconnectController.hasStartedConnectionAttempt()) {
+      this.reconnectController.startedConnectionAttempt(false);
+    } else {
+      this.reconnectController.startedConnectionAttempt(true);
+    }
     // reconnect needs to re-resolve endpoint url, which will also refresh credentials on client if they are expired
     if (reconnecting || endpointUrl === undefined) {
       try {
         if (this.configuration.chimeClient.getMessagingSessionEndpoint instanceof Function) {
-          endpointUrl = (await this.configuration.chimeClient.getMessagingSessionEndpoint())
-            .Endpoint.Url;
+          const response = await this.configuration.chimeClient.getMessagingSessionEndpoint();
+          // Check for aws sdk v3 with v2 style compatibility first
+          if (response.Endpoint?.Url) {
+            endpointUrl = response.Endpoint.Url;
+          } else {
+            // Make aws sdk v2 call
+            const endpoint = await this.configuration.chimeClient
+              .getMessagingSessionEndpoint()
+              .promise();
+            endpointUrl = endpoint.Endpoint.Url;
+          }
         } else {
           endpointUrl = (
             await this.configuration.chimeClient.send(new GetMessagingSessionEndpointCommand({}))